Refactor crypto integration with arrow
arrow >= 20.0.0 refactored the crypto API to support better
multithreading; align to new API with backwards compatibility.
